+++ test/e2e/support/utils/are-pre-publish-checks-enabled.js
@@ -1,3 +1,7 @@
+/**
+ * Verifies if publish checks are enabled.
+ * @return {boolean} Boolean which represents the state of prepublish checks.
+ */
 export async function arePrePublishChecksEnabled() {
 	return page.evaluate( () =>
 		window.wp.data.select( 'core/editor' ).isPublishSidebarEnabled()