+++ packages/editor/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 9.0.3 (Unreleased)
+
+### Bug Fixes
+
+- `getEditedPostAttribute` now correctly returns the merged result of edits as a partial change when given `'meta'` as the `attributeName`.
+
+## 9.0.2 (2018-11-22)
+
+## 9.0.1 (2018-11-21)
+
+## 9.0.0 (2018-11-20)
+
+### Breaking Changes
+
+- `PostPublishPanelToggle` has been removed. Use `PostPublishButton` instead.
+
 ## 8.0.0 (2018-11-15)
 
 ### Breaking Changes
@@ -40,7 +56,7 @@
 
 ### Deprecations
 
-- `wp.editor.PostPublishPanelToggle` has been deprecated in favor of `wp.editor.PostPublishButton`.
+- `PostPublishPanelToggle` has been deprecated in favor of `PostPublishButton`.
 
 ### Polish
 
@@ -106,7 +122,7 @@
 
 ### Deprecations
 
-- `wp.editor.PanelColor` has been deprecated in favor of `wp.editor.PanelColorSettings`.
+- `PanelColor` has been deprecated in favor of `PanelColorSettings`.
 
 ### New Features
 
@@ -123,7 +139,7 @@
 - The `value` property in color objects passed by `withColors` has been removed. Use `color` property instead.
 - `RichText` `getSettings` prop has been removed. The `unstableGetSettings` prop is available if continued use is required. Unstable APIs are strongly discouraged to be used, and are subject to removal without notice, even as part of a minor release.
 - `RichText` `onSetup` prop has been removed. The `unstableOnSetup` prop is available if continued use is required. Unstable APIs are strongly discouraged to be used, and are subject to removal without notice, even as part of a minor release.
-- `wp.editor.RichTextProvider` has been removed. Please use `wp.data.select( 'core/editor' )` methods instead.
+- `RichTextProvider` has been removed. Please use `wp.data.select( 'core/editor' )` methods instead.
 
 ### Deprecations
 
@@ -151,12 +167,12 @@
 - `getSharedBlocks` selector has been removed. Use `getReusableBlocks` instead.
 - `editorMediaUpload` has been removed. Use `mediaUpload` instead.
 - Change how required built-ins are polyfilled with Babel 7 ([#9171](https://github.com/WordPress/gutenberg/pull/9171)). If you're using an environment that has limited or no support for ES2015+ such as lower versions of IE then using [core-js](https://github.com/zloirock/core-js) or [@babel/polyfill](https://babeljs.io/docs/en/next/babel-polyfill) will add support for these methods.
-- `wp.editor.DocumentTitle` component has been removed.
+- `DocumentTitle` component has been removed.
 - `getDocumentTitle` selector (`core/editor`) has been removed.
 
 ### Deprecations
 
-- `wp.editor.RichTextProvider` flagged for deprecation. Please use `wp.data.select( 'core/editor' )` methods instead.
+- `RichTextProvider` flagged for deprecation. Please use `wp.data.select( 'core/editor' )` methods instead.
 
 ### Bug Fixes